Mass rename of the library
