+++ library.properties
@@ -1,9 +1,9 @@
-name=WiFi101OTA
-version=1.0.2
+name=Arduino_WiFiOTA
+version=1.1.0
 author=Arduino
 maintainer=Arduino <info@arduino.cc>
 sentence=Update sketches on your board over WiFi
 paragraph=Requires an Arduino/Genuino SAMD board
 category=Other
 url=http://www.arduino.cc/en/Reference/WiFi101OTA
-architectures=samd
+architectures=samd,megaavr
